Split receiving and handling of requests
